pyproject.toml
[tool.poetry]
name = "langchain-notes"
version = "0.1.0"
description = ""
authors = ["BikS <giorgos.marinos@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
langchain = "^0.3.14"
langchain-core = "^0.3.29"
langchain-openai = "^0.2.14"
python-dotenv = "^1.0.1"
langchain-anthropic = "^0.3.1"
langchain-google-genai = "^2.0.8"
langchain-community = "^0.3.14"
chroma = "^0.2.0"
beautifulsoup4 = "^4.12.3"
html2text = "^2024.2.26"
langchain-chroma = "^0.2.0"
chromadb = "^0.5.0"
langchain-google-community = "^2.0.4"
networkx = "^3.4.2"
spacy = "^3.8.3"
sentence-transformers = "^3.3.1"
faiss-cpu = "^1.9.0.post1"
pypdf = "^5.1.0"
reportlab = "^4.2.5"
voyageai = "^0.3.2"
matplotlib = "^3.10.0"
jsonschema = "^4.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
